Unix: Deploy plugins to (libdir)/doomsday-plugins
On (non-Mac) Unix platforms, the plugins are now installed in a special
directory reserved for them. 'libdir' in the paths config file can be
used to change the directory. The regular shared libs are installed to
the main libdir (e.g., /usr/lib).

Fixed detection of shared libraries in de::LibraryFile so that the
".so.x.y.z" symbolic links aren't loaded as separate plugins.