Refactor|FS1: Avoid manipulating the open-files list in accessFile()
Restructed logic for interpreted file instantiation so as to avoid
manipulating the open-files list when answering is-file-accessible
queries. In turn this means that one fewer DFile handle is needed,
further improving startup performance.

Rather than returning a handle to the opened file (that which was
added to the opened-file list), tryOpenFile() now returns the file
itself (leaving it to the caller to decide whether to list and/or
index it later if needed).