Refactor|FS1: Made the Wad LumpIndexes of FS1 publically visible
Rather than wrapping access to the internal indexes, make these
available (via const reference) to users of the FS1 class. This
avoids the needless API duplication and call routing previously
done by the FS1 module.

The only method of FS1 which takes a lump number as argument is
nameIndexForLump(), which given an absolute index, returns the
applicable LumpIndex (and translates the lump number into range).