Mac OS X: Deploy FMOD Ex library to the Doomsday.app Frameworks folder
If the FMOD Ex library is found inside the audio_fmod bundle, Doomsday
will attempt to open it as a Doomsday plugin. This leads to an
unnecessary error. It is better to deploy the FMOD Ex shared library
with the other shared libs in Doomsday.app/Contents/Frameworks.

Note: This expects that "engine" is built before the plugins, as the
library is copied directly to the built Frameworks folder.