Fixed: Log message argument issues
Note that "%lu" is NOT a valid de::Log entry argument format. All
integer values are converted to 64-bit integers when the entry is
created. These were replaced with "%u".

Added the missing "%c" format for printing one Unicode character. See
de::String::patternFormat().

Fixed a compilation issue due to the "long int" not having a constructor
in LogEntry::Arg.

Todo: Check that the constructors aren't conflicting on a 32-bit
compiler.