FluidSynth|Unix: Use the system's libfluidsynth
Instead of the embedded libfluidsynth under "external" in the deng
source tree, use pkg-config to build with libfluidsynth installed
on the system.

On Mac OS X, the embedded FluidSynth is the default. On other UNIXes,
the system libfluidsynth is the default.

The configuration option is "deng_embedfluidsynth" (default: not set).