Refactor|PathTree: All path arguments given to PathTree now use de::Uri
Rather than accept path arguments as String, the PathTree API now
uses Uri and paths are built using it's built in path name map.
In turn it is now unnecessary to specify a delimiter at insertion
time as all URI's use a forward slash (delimiter translation is
handled at de::Uri level).

Note that URIs used to populate a PathTree are NOT resolved. This
means that any existing symbolics in their paths will thusly exist
in the PathTree also. Presently this is not relevant because trees
are only ever built from resolved resource search paths.