Refactor|ResourceNamespace: ResourceNamespace no longer piggyback's Uri::hashPathNodeName()
A resource name is NOT the same as a PathTree and/or Uri leaf name.
The reason being that the former does not include a file extension
and because there may be multiple names for a resource - (none of
which are required to match the associated file name).

ResourceNamespace now implements its own name hashing algorithm.
Uri's name hashing algorithm is now encapsulated by Uri::PathNode.

Todo: ResourceNamespace's custom NameHash class is now unnecessary
and could be reimplemented with (say) QMultiHash.