Moved UI-related functionality to the "ui" source subfolder
The term "user interface" here encompasses everything related to the
user interface: windows, display mode, controls, bindings, input
devices, and the engine UI. A (dedicated) server will have no use for
any of this.

fi_main/finaleinterpreter is a bit of a grey area, but they were
included in the UI as well. InFine is a UI animation system, so it
should be something that dedicated servers can (mostly) ignore.