Fixed|Unix: Expanding ~ in the iwaddir
The paths set with "iwaddir" from UnixInfo or with environment variables
were not being expanded, causing the engine to fail locating any IWADs.

Also added Uri::fromNativeDirPath() for telling Uri that the referenced
path points to a directory rather than a file, as the rest of FS1 uses
this distinction, while de::NativePath (from libdeng2) does not.