Use fixed-precision math in bounding box vs. line side checks
The original DOOM used fixed-precision math in all playsim math. When
Doomsday switched to floating point values, the behavior of some
collision tests was altered. This commit attempts to address this by
restoring the original fixed-precision math in the specific case of
testing on which side of a line a bounding box is located.