libdeng2: Don't override build's default log entry level
We are currently using the practice of basing the default log entry
level on the build type. However, de::App was overriding this with
the level in Config. This would allow setting a persistent log level
and changing it at runtime, but since the user has no way of modifying
the Config at present, we'll be just using the build type based
level for now.