Refactor: Removed the lump index concept from File1
This concept was implemented at this level as an interim measure, as
it enabled a straightforward mechanism to achieve a single interface
to the lump indexes of Zip and Wad files before File1 was redesigned
to use polymorphism.

Now, this concept would be better implemented as a specialisation of
File1 which defines the interface mechanisms for a segmented file.