Refactor: Uri's "path nodes" renamed to "segments" as per URI spec
The "path nodes" of the Uri had nothing to do with PathTree::Nodes or
with tree nodes in general. They simply mark portions of the URI's
path so that the start and end location of each segment of the URI path
are known.

Refactored the class implementation to simplify the code and remove
redundancies. The manually managed linked list of segments was replaced
with a QList.

Updated the unit test for the Uri class.