Optimize: Restructured model loading
The old model loading mechanism was rather awkward and suboptimal
as it required an instance of the runtime representation of a model.
Therefore regardless whether the resource file was even found it
was necessary to first allocate this object.

Also split the logic with discrete search, recognise and interpret
steps in preparation for further refactorings.