Refactor|libdeng2: Separated ZipArchive from the generic Archive class
The generic Archive class is now agnostic of the serialization format.
It handles the management of a set of data blocks that are identified
by paths, and handles the caching of the data.

The ZipArchive is only responsible for reading the ZIP format,
populating the archive with suitable entries, and writing the contents
of the archive using a Writer.