Refactor|libdeng2: Improved API; adhered to libdeng2 conventions
Never force the user to rely only on unsafe API methods. If there is a
benefit to be gained (such as, potentially, performance in the case
of StringPool::stringRef()), it should be offered as an alternative
to the safe (and potentially slower) methods.

The safe method should be default/shorter-named one.

The unsafe aspects of the methods need to be clearly explained in the
apidoc.

Removed unneeded complexity from certain methods (functionality that
was not taken advantage of anywhere in the code base).

Added StringPool::InvalidIdError.