Fixed|libdeng2: Files can only be immutable input streams
The assumption with modifiable input streams is that the read bytes
are gone after having been read. This assumption is not compatible
with how byte array files work: the read bytes persist. Therefore,
byte array files can only be used as immutable streams.

However, Reader will automatically check if the stream can be casted
to a byte array, as byte arrays provide random access and are
therefore more useful for reading.