libdeng2: Renamed ArchiveFile to ArchiveEntryFile, added PackageFolder
The name "ArchiveFile" was not consistent with the intended naming
convention: it does not represent an Archive, only a single entry in
an Archive. Thus it was renamed to ArchiveEntryFile.

Added PackageFolder that will be used instead of Folder to represent
the root folder of a ZIP archive. It provides a more convenient way
to access the contained Archive.