libdeng2: Added TimeValue, fixed bugs in de::Time
TimeValue allows manipulating time in expressions and lets variables
hold a Time as a value.

There were some bugs in de::Time related to time arithmetic and asText
formatting.

The new built-in functions "Time" and "timedelta" can be used to
create time values and to calculate the difference between two times.

Updated the kitchen sink test script for trying out the new
functionality.