Refactor|API: Replaced DD_MaterialForTextureUniqueId with DD_MaterialForTextureUri
The public API function DD_MaterialForTextureUniqueId was necessary
prior to the standardization of the texture collection subspaces and
the introduction of URIs.

Textures can now be looked up using a de::Uri of either "regular" URI
or absolute URN form, thus there is no need for the API to make this
internal indexing interface public.