libdeng2: Describing files in a human-friendly manner
Because the libdeng2 FS is a fully virtual file system, the user needs
to know exactly where the data comes from in order to understand the
system and how actions on source data affect the files present in the
libdeng2 FS.

Added a new method File::description() that returns a full, human
friendly description of the file. This includes possible read-only
status, possible path in the FS tree (if in the tree), type of the
file, the possible source, the originating feed, and any other relevant
information specific to the file's type (e.g., type of shared library
if the library is loaded in memory).

File::description() should be used in high-level log messages to give
the user full contextual information about a file.

Low-level/debug messages should stick to name() and path() if possible
to avoid flooding the log with redundant information.