Optimize: Removed the redundant central texture index
As there is no longer a need to reference textures with a contiguous
range of indices, the central texture index itself was redundant.

This is because the texture manifests persist engine resets and all
bookkeeping is fully dynamic.

The unused central texture index and the associated textureid_t have
now been removed.