FMOD: Sound effects are not resampled to 'sound-rate'
FMOD's output is by default played at the sound card's configured rate.
Typically this is around 44100 Hz (or 48000 Hz).

FMOD will resample audio on the fly if needed, so there is no need
for Doomsday to get involved with its poor linear-interpolating
upsampler.

Added a new SFX interface property that informs the engine's sfx module
that upsampling is not required.