libdeng2: Writer and Reader support versioning
Going forward, it is crucial for backwards compatibility that when
serializations of various objects change, the readers can adapt and
read the old version instead of expecting the latest.

By default Writer and Reader assume that the latest version is being
used. When writing at the stream or container/file level, one should
use the withHeader() methods to include/read the serialization protocol
header.