Refactor|Patch: Moved paletted image composition out of Patch
The potentially complicated task of compositing two or more images
should be performed in a way which is agnostic of the serialization
formats of the images being composited.

Paletted image composition is now handled by compositePaletted() in
gl_texmanager.cpp.

Todo: Optimize the currently naive compositePaletted() once extended
and generalized with more of the functionality one would expect from
such a routine.