Fixed: Don't assume a sprite is in Patch format
At sprite definition time validate whether the associated resource
is a Patch and if so, obtain the world dimension and offset values
to attribute with the sprite. Otherwise, ignore the resource for
now and allow the texture loader to attribute the world dimensions
from the pixel dimensions of image resource at load time.