Patch: Clamp post pixel ranges in Patch::load()
The height dimension value as declared in the patch header may well
be less than the "real" height of the composited image. Consequently
it is necessary to clamp post pixel ranges when compositing the image
to the output buffer.

This is the reason why map drawing in the id tech 1 software renderer
can be seen to "overdraw" posts - the wall column drawer is working
with post pixel ranges rather than the "logical" height declared in
the header.

Todo: Apply this information to texture/material creation.