Fixed|CompositeTexture: Dynamically enlarge CompositeTexture geometry
The behavior of the id Tech 1 software renderer's column drawer is
to ignore the logical dimensions of composite textures. Instead a
wall column is drawn with respect to the actual pixel dimensions of
the component patches.

When preparing CompositeTextures the final geometry of the composited
texture is calculated according to the union of all component patch
geometries. Which, will in turn be propagated to derived materials.

Fixes: Bug#1071 (Patch texture composition logic errors)