DehReader: Fixed (parser bug) skipping over text patches
After parsing a text patch the parser should skip to the start of
the next line rather than to start of the next section.

Text patches are parsed with different semantics, they are not a
"section", so the skip-to-next-section logic (which assumes that
the parser is currently reading a section) won't work as expected.