libdeng2: Clear log buffer before unloading any library
It is possible that the library has created log entries that are still
in the log buffer. Previously this was done in 'engine', but now since
libdeng2 is in control of all plugin loading and unloading, it can
clear the log buffer in one central place (just before closing the
library).