Fixed|Doom: Secret exit switch coords in DOOM E3M6
This mapping construct triggered a compatibility logic for the
id Tech 1 column renderer when generating geometry, which resulted
in the texture coords being miscalculated. An extra qualification
check was added to the logic to ensure it is only applied when the
back floor plane is above the front ceiling plane.