Fixed|Server|Windows: Compilation of the server
Added missing dummies and adjusted Windows-specific routines for
a SERVER build.

The linker warning about the wrong name in doomsday.def should be ignored:
for the time being, identical exports must be used for doomsday.exe and
doomsday-server.exe, and using the same .def file is the best way to
ensure that.