qmake: No more exporting symbols out of the executable
Adjusted the build settings in the qmake project files so that no more
symbols are exported out of the executables. This has been a long-time
annoyance and a kludge that is now finally being removed. In the future,
exports will be handled manually in a way that allows plugins to work
with any executable (client or server).