Light Decorations: Decorations are now owned by the material
As material animations are now modelled using layer stages, light
decorations also need to be handled a bit differently. This first
phase is simply about moving this data into material_t, the next
step is to extend the implementation to add a stage animation
mechanism for the decorations.

Todo: Reimplement support for the external/custom decoration def
selector flags.