Refactor|DMU: Dummies implemented using polymorphism
The dummies are no longer stored in three arrays, but instead a single
QSet. Thanks to de::MapElement's polymorphism the dummies can be
handled in a more abstract fashion.

Also noteworthy: the DMU API version is not checked any more. API
versioning is now taken care of elsewhere for all public APIs.