Refactor: Moved Texture variant selection to Texture::chooseVariant()
Management of the pooled texture variant specifications should be
done within the Textures collection, as the variant instances are
also logical objects.

Doing this within the GL texture manager unnecessarily ties the
resource domain to the render domain.

However, this is out of scope for this work branch so I won't be
taking this any further.