Refactor: Relocated TextureVariantSpec_Compare() to texturevariant.cpp
Although some of the specification properties are only meaningful
for the purposes of rendering, which is a client only concern, the
server presently still needs to manage context specialized variants
at a logical level.

However, ultimately there should be no need for Texture::Variant or
their specification on server side.