Shell: Set up framework for curses-based shell app
Started work on building a shell app. The UI is based on the same
libdeng2 UI widgets that will act as the foundation of the client's
graphics UI.

Added a curses-specific TextCanvas, a widget class for widgets drawn
on a TextCanvas, a RootWidget class that owns a TextCanvas, LogWidget
for the server's message history, and ShellApp that will set up a shell
connection and create the UI widgets.