Refactor|libdeng2: Log entry formatting and output
Added the LogSink class that represents a destination for flushing
log entries to.

LogSink::IFormatter is an interface that converts log entries to one
or more lines of text. A LogSink is not required to use a formatter.

The previous hard-coded formatting was moved to
MonospaceLogSinkFormatter, which currently uses a hard-coded line width.
The built-in log sinks all use their own monospace formatters.