Fixed|Network: Writing multiple messages simultaneously
In certain situations, a network message is written during the writing
of another message (e.g., a console message being echoed to a logged
in client during the composing of a player fix message).

Now ongoing writers are pushed onto a stack automatically so that the
earlier message can be resumed once the new message is done.

Messages are consequently no longer composed directly in the netbuffer,
but rather in each Writer instance's own dynamic buffer.