Client: Added ServerLink (derived from shell::AbstractLink)
ServerLink contains the functionality that sys_network + protocol used
to have: discovering servers on the network and managing the connection
to one of them server.

shell::ServerFinder was incorporated to automatically locate servers
on the local network. They are included in the list of found servers
automatically.

Todo for later: Revise the client GUI for connecting to a server
(when the new UI framework is up and running).