Windows|Fixed: Application refresh loop behavior, LegacyWidget/BusyWidget issues
It seems that on Windows the only way to run the app refresh loop is to
use a QTimer (and not singleshot calls). Also corrected issues related to busy
widget drawing/update during transitions between normal and busy mode. View
resizing is also ignored if the engine is shutting down.