Refactor: Collate BSP elements with an iterative pre-order algorithm
It is no longer necessary to implement BSP map element collation
with a post-order algorithm (which was implemented recursively)
because the BspBuilder defers deletion of it's guiding tree.

Obviously this change affects the order of the elements within
GameMap's various LUTs, however, this is irrelevant as there is
no longer any logic dependent on this order.