Fixed|Definitions: Appending layer stages to auto-generated materials
The unlikely but possible case of the mod directive being used to
append additional layer stages to an auto-generated material def
was undefined.

Presently the most logical course of action in this situation is to
copy the property values from the previous stage at parse time.

Todo for later: Ideally the stage animation mechanism would handle
this with its own logic, whereby if a property value is undefined
for the current stage; take the last known value from an earlier
stage.