libcommon: Added cvar 'game-pause-mapstart-tics'
In some games (like libdoom) it is important to keep game time
paused while the engine's busy transition effect is ongoing, or
otherwise the player's view would be obscured during the first tics
of a map.

In libdoom, the default for game-pause-mapstart-tics is -1, which
means the forced pause is equal to the engine's con-transition-tics.

In libheretic and libhexen, the default is 7 tics (0.2 seconds),
because the default transition is a crossfade that does not obscure
the view for very long. (Plus it makes for a neat visual effect.)