Documentation: Console commands and variables as Amethyst sources
Generated an .ame source for each console command and variable based
on the cphelp/conhelp files. These are intended as the new master
copy of this documentation. Later on the console help files will be
generated out of these sources.

Todo: Script for generating categorized wiki pages for commands and
variables, plus index tables.