Documentation|Console|All Games: Updated cphelp.txt and conhelp.txt files
The console and control panel help strings are now generated out of
the Amethyst sources under "doomsday/doc".

At this point this is more of a proof-of-concept that demonstrates
how Amethyst is used for documentation that appears both in-game and
in the wiki. Most of the commands and variables presently still use
unformatted content.

Todo for later: Apply more Amethyst formatting, improve in-game output.