Refactor|GameMap: Construct the BSP leaf blockmap as soon as the BSP is built
The BSP leaf blockmap would be useful to accelerate other tasks in
the map load process, so get this ready at the earliest opportunity.

Also added back the DE_API_MAP_v1 definition in apis.h, as it may be
useful in future for reference.

Todo for later: Optimize S_DetermineBspLeafsAffectingSectorReverb()
(and others), making use of the BSP leaf blockmap.