Refactor|LineDef: Redefined LineDef's responsibilities somewhat
Inlined openRange() and visOpenRange() as these do not belong in
LineDef itself however it is convenient to have them there. Although
this functionality appears to be suitable for inclusion in LineDef,
it would be altogether too easy to inadvertently subvert the logic
of R_OpenRange(), on which so much depends (geometry construction,
vanilla playsim logic emulation, etc...). Hopefully a better home
and/or mechanism will present itself later.

configureTraceOpening() was extracted and merged into GameMap's
setTraceOpening() opening for now (it shouldn't fall to GameMap to
update the state of LOS traces).