Refactor|libdeng2: Improved Config's public interface
There are now set() methods that automatically create variables and
records in the Config namespace, so that the Config script is not
solely responsible for making sure they exist before the values are
accessed. Also added a [] operator for Config itself and made the
get*() methods 'const'.

Improves robustness of the Config mechanism: consider downgrading to
an old version where the Config is different…