Sector: Redefined the aabox as the union of the BSP leaf's aaboxs
As a sector in Doomsday has no geometry of its own, the aabox (and
thus the rough area approximation) should be taken from BSP leafs
instead.

This also means that the rough area approximation for a sector is
now derived from the BSP leafs, addressing a shortcoming in the
space volume calculation done to determine the environmental audio
characteristics of the sector.