Refactor: Moved sound emitter from Surface to Plane and LineDef::Side
Surface has no geometry of it's own thus it has no origin in the map.

Moving the sound emitters out of Surface and upwards to Plane and
LineDef::Side clarifies the surface abstraction and in the process
removed some ugly logic where the surface made decisions according
to the type of it's owner.