Refactor: Renamed LineDef as Line
The name "LineDef" leads one to think that this is definition for a
would-be map line rather than the line itself. LineDef only makes
sense in the context of the map format interpretor, when lines are
produced from LINEDEFS.

As we are breaking API compatibility for this release in any case,
now is the time to effect the rename to "Line".