Refactor: Dumped the redundant SideDef class
The opaque sidedef_s used by the public API now maps to Line::Side
within the engine. DMU has been rewired accordingly. Publicly there
is no apparent difference.

Internally, as all Lines now have two Sides, maintaining a separate
LUT for the sides is unnecessary. Instead, GameMap provides a means
to translate to/from indices and pointers to Side. Accordingly, the
index of a Side is now equal to that of the owning Line except that
the high bit is set to mark a "back" side (assumes little-endian).

Todo for later: Update the games' use of DMU so that they too can
reference the logical sides of a Line.