Refactor: Renamed SideDef as Side
The name "SideDef" leads one to think that this is a definition for
a would-be map line side rather than the side itself. SideDef only
makes sense in the context of the map format interpretor, when line
sides are produced from SIDEDEFS.

As we are breaking API compatibility for this release in any case,
now is the time to effect the rename to "Side".