Fixed|Weapon Info: "Static" psprites (weapon offset scale)
When drawing psprites, the weapon offset scale that is used for
preventing the lowering of the psprite was not applied as intended:
WEAPONTOP cannot be ignored as that is the highest offset that the
psprite will reach (non-zero).

Todo for later: Make WEAPONTOP a DD_value so that it doesn't have to
hardcoded.