Fixed: Obtain the Sector for a given half-edge from it's owning BspLeaf
The pointer in HEdge does not always reference the correct Sector as
this is presently determined by errant logic in the BSP partitioner.
Instead we will redirect this query to the owning BspLeaf as this is
the final authority as to which Sector the half-edge is "in".

The now build time only sector pointer in HEdge was relocated to the
Partitioner's HEdgeInfo.

Todo: Address the cause of this issue in the partitioner.