DMU API: Added property DMU_ARCHIVE_INDEX
The archive index is the position of the relevant data and/or
definition for the map element in the "archived" map. The archive
in this context is original source of the map element (e.g., from
an id Tech 1 format map).

For example, in the case of a DMU_SIDE element that is produced
from an id tech 1 format map, the archive index will be the index
of the definition in the SIDEDEFS data lump.

Todo: This info should be used when translating legacy indexes,
such as those used in the Doom ver 1.9 save game reader.