libgui: Implemented image/texture atlases
Added Atlas, RowAtlasAllocator and AtlasTexture.

de::Atlas is an abstract class and provides the primary interface for
image-based atlases. It optionally maintains a backing store.

de::Atlas::IAllocator is the interface for an atlas allocator that
decides where content is placed inside the atlas.

de::RowAtlasAllocator is a row-based allocator for content whose
heights are fairly constant (e.g., text strings).

de::AtlasTexture is a concrete atlas class that uses a GLTexture as
the physical storage medium.