Refactor|Partitioner: Sort partition intercepts at merge time
Analysis shows that always maintaining the order of the intercepts
has near zero time/complexity profit when compared to sorting them
at merge time.

The partition intercept list is now modeled with a QList, with the
sort and merge phase implemented in-place.

Looking forward, this revised design makes more sense if we decide
to multithread the node build process.