Refactor|BSP Builder: Began remodelling the space partitioning logic
Returning to this code after the enhancements made for Doomsday 1.10
it is clear that the high level design of this module was somewhat
compromised in the process.

Before going further with the extraction of the geometry generation
from the map renderer, it makes sense to clarify the implementation
once more.