DMU API: Renamed DMU_BASE; removed DMU_LINE_COUNT, DMU_HEDGE_COUNT
DMU_EMITTER is a better name for this property as it references the
sub-component of the relevant map element used for emitting sounds.

The presently unused DMU_LINE_COUNT and DMU_HEDGE_COUNT property
identifiers were removed as these set the wrong precedent for how
the linked objects of a map element should be enumerated.

If in future we have call to enumerate the linked objects of some
map element, this functionality should be implemented as a runtime
query (e.g., numlines = P_Countp(sectorPtr, DMU_LINE)).