SectionEdge: Improvements
Amalgamated WallDivs functionality and switched to index based
iteration mechanism for intercepts.

It is now clear there that there is no need for a WallDivs-like
further abstraction of this component, given the presence of a
half-edge data structure (accelerating the intercept search phase)
and the possibility to defer sorting.

Dumped the old WallDivs component.