Fixed|FakeRadio: Plane shadows drawn in map hack sectors with zero height
A fairly uncommon situation which involves the use of an unclosed
sector in combination with a "mid grate" hack. With such a construct
the shadow lines are presently linked to the BSP leaf in the "outer"
sector, however it is necessary to check that the actual sector on
the front side (taken from the left most half-edge) has a positive
world volume at draw time.